Adicionando Data e Wrapper da ESPN
