--- configs.py
@@ -1,92 +1,0 @@
-agent = {
-  'User-Agent': 'Mozilla/5.0 (Windows NT 6.3; Win64; x64; Trident/7.0; rv:11.0) like Gecko'
-}
-
-heuristic = [
-  {
-    'key': 'jogador',
-    'weight': 9
-  },
-  {
-    'key': 'player',
-    'weight': 8
-  },
-  {
-    'key': 'futebol',
-    'weight': 7
-  },
-  {
-    'key': 'status',
-    'weight': 6
-  },
-  {
-    'key': 'stats',
-    'weight': 6
-  },
-  {
-    'key': 'soccer',
-    'weight': 6
-  },
-  {
-    'key': 'football',
-    'weight': 5
-  }, 
-  {
-    'key': 'perfil',
-    'weight': 5
-  },
-  {
-    'key': 'profile',
-    'weight': 4
-  },
-  {
-    'key': 'people',
-    'weight': 3
-  },
-  {
-    'key': 'basketball',
-    'weight': -4
-  },
-  {
-    'key': 'volleyball',
-    'weight': -4
-  },
-  {
-    'key': 'cricket',
-    'weight': -4
-  },
-  {
-    'key': 'rugby',
-    'weight': -4
-  },
-  {
-    'key': 'handball',
-    'weight': -4
-  },
-  {
-    'key': 'floorball',
-    'weight': -4
-  },
-  {
-    'key': 'water',
-    'weight': -4
-  },
-  {
-    'key': 'handball',
-    'weight': -4
-  },
-  
-]
-
-links = [
-  'https://globalsportsarchive.com',
-  'https://www.playmakerstats.com',
-  'https://fbref.com',
-  'https://br.soccerway.com',
-  'https://www.goal.com/br',
-  'https://www.ogol.com.br',
-  'https://www.transfermarkt.com.br',
-  'https://www.sambafoot.com/br',
-  'https://www.foxsports.com',
-  'https://www.espn.com.br'
-]
