--- out.py
@@ -1,1773 +1,0 @@
-total_harvest_ratio = 0.114886731
-
-crawledLinks = {
-  'https://www.ogol.com.br':
-  {
-    'visited_links': 98,
-    'harvest_ratio': 0.02040816326530612,
-    'relevant_links': 2,
-    'thread': 'Thread-11',
-    'frontier': [
-      {
-        'weight': 60.0,
-        'link': 'https://www.soccerzz.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://www.playmakerstats.com'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.ogol.com.br'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.facebook.com/ogolcombr'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.ogol.com.br/index.php?&redirm=1'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.calciozz.it'
-      },
-      {
-        'weight': 0,
-        'link': 'http://refpa.top/L?tag=d_27572m_97c_&site=27572&ad=97'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.leballonrond.fr'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.instagram.com/ogol_oficial'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.fussballzz.de'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.zerozero.pt'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.ogol.com.br/text.php?id=11739'
-      },
-      {
-        'weight': 0,
-        'link': 'https://twitter.com/ogolcombr'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.ceroacero.es'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.youtube.com/channel/UCjWcRCWje4HvkD63qqIM8JQ'
-      },
-      {
-        'weight': 0,
-        'link': 'https://facebook.com/ogolcombr'
-      }
-    ]
-  },
-  'https://www.espn.com.br':
-  {
-    'visited_links': 0,
-    'harvest_ratio': 0,
-    'relevant_links': 0,
-    'thread': 'Thread-19',
-    'frontier': []
-  },
-  'https://www.foxsports.com':
-  {
-    'visited_links': 25,
-    'harvest_ratio': 0.0,
-    'relevant_links': 0,
-    'thread': 'Thread-17',
-    'frontier': [
-      {
-        'weight': 0,
-        'link': 'https://www.foxcorporation.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://help.fox.com/foxsports'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.youtube.com/foxsports'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsports.com/regions'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsports.com/presspass/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.instagram.com/foxsports'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsports.com/privacy-policy'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxdeportes.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsports.com/terms-of-use'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsports.com/privacy-policy#relevant_advertising'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.fox.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsports.com/sitemap'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.facebook.com/foxsports'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.twitter.com/foxsports'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsuper6.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://advertising.foxsports.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsports.com/supports/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxnews.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsports.com/do-not-sell'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.caffeine.tv/foxsports'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxcincy.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxbet.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxcareers.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.foxsports.com/rss-feeds'
-      }
-    ]
-  },
-  'https://www.goal.com/br':
-  {
-    'visited_links': 11,
-    'harvest_ratio': 0.09090909090909091,
-    'relevant_links': 1,
-    'thread': 'Thread-9',
-    'frontier': [
-      {
-        'weight': 50.0,
-        'link': 'https://www.linkedin.com/company/footballco/jobs/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://play.google.com/store/apps/details?id=com.freerange360.mpp.GOAL'
-      },
-      {
-        'weight': 0,
-        'link': 'https://twitter.com/GoalBR'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.facebook.com/GoalBR'
-      },
-      {
-        'weight': 0,
-        'link': 'https://go.onelink.me/2947384588/479c5d2f'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.instagram.com/goalbrasil'
-      },
-      {
-        'weight': 0,
-        'link': 'https://itunes.apple.com/app/goal.com/id518026818'
-      }
-    ]
-  },
-  'https://globalsportsarchive.com':
-  {
-    'visited_links': 143,
-    'harvest_ratio': 0.13986013986013987,
-    'relevant_links': 20,
-    'thread': 'Thread-1',
-    'frontier': [
-      {
-        'weight': 140.0,
-        'link': 'https://globalsportsarchive.com/players_abroad/soccer/'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://globalsportsarchive.com/competitions/soccer/'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://globalsportsarchive.com/competition/soccer/ligue-1-uber-eats-2020-2021/regular-season/46313/'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://globalsportsarchive.com/competition/soccer/serie-a-tim-2020-2021/regular-season/48277/'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://globalsportsarchive.com/competition/soccer/laliga-santander-2020-2021/regular-season/48255/'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://globalsportsarchive.com/competition/soccer/premier-league-2020-2021/regular-season/47188/'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://globalsportsarchive.com/competition/soccer/uefa-champions-league-2020-2021/round-of-16/47146/'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://globalsportsarchive.com/competition/soccer/bundesliga-2020-2021/regular-season/46282/'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://globalsportsarchive.com/sport/soccer/-/'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://globalsportsarchive.com/competition/soccer/copa-de-la-liga-profesional-2021/group-stage/51615/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/competition/australian_football/toyota-afl-premiership-2021/regular-season/50587/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/sport/american_football/-/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/sport/american_football/2021-02-08/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/competition/american_football/cfl-2019/grey-cup/34893/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/sport/australian_football/-/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/competitions/australian_football/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/competition/american_football/ncaa-division-i-fbs-2020-2021/bowls/50727/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/sport/australian_football/2021-03-25/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/competitions/american_football/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://globalsportsarchive.com/competition/american_football/nfl-2020-2021/nfc-divisional-play-offs/43538/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/cricket/kfc-big-bash-league-2020-2021/regular-season/47426/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/kabaddi/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.instagram.com/globalsportsarchive/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/softball/2021-07-17/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/rugby/guinness-pro14-2020-2021/regular-season/48895/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.facebook.com/globalsportsarchive/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/field_hockey/fih-pro-league-2020/regular-season/37814/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/ice_hockey/khl-2020-2021/conference-semi-finals/47787/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.datasportsgroup.com'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/counterstrikego/2021-03-24/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/alpine_skiing/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/rainbowsix/2021-01-23/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/badminton/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/cricket/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/water_polo/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/handball/liqui-moly-handball-bundesliga-2020-2021/regular-season/48326/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/counterstrikego/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/counterstrikego/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/baseball/cpbl-2020/taiwan-series/42772/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/field_hockey/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/field_hockey/odisha-hockey-world-cup-2018-india/final/24198/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/ice_hockey/iihf-world-championship-2019-slovakia/final/29926/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/ice_hockey/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/basketball/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/nba2k/2020-08-29/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/basketball/basketball-champions-league-2020-2021/play-off-round/46849/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/water_polo/womens-european-water-polo-championship-2020-budapest/final/36930/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/volleyball/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/cricket/womens-twenty20-international-2021/regular-round/51867/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/motorsports/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/water_polo/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/baseball/npb-2020/japan-series/44251/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/rugby/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/handball/delo-womens-ehf-champions-league-2020-2021/group-stage/46089/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/floorball/2019-09-22/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/baseball/mlb-2020/world-series/39655/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/handball/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/cricket/icc-odi-championship-2020/regular-round/37451/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/volleyball/fivb-volleyball-mens-nations-league-2019/final/32195/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/rugby/super-rugby-2019/final/34878/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/basketball/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/baseball/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/water_polo/len-champions-league-2020-2021/group-stage/45591/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/tournament/golf/world-golf-championships-dell-technologies-match-play-2021/46755/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/ice_hockey/khl-2020-2021/regular-season/47785/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/judo/2019-11-24/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/basketball/nba-2020-2021/regular-season/50493/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/ice_hockey/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/basketball/turkish-airlines-euroleague-2020-2021/regular-season/46433/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://discord.gg/nYyvYNvN'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/cricket/icc-test-championship-2020/regular-round/38274/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/volleyball/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/baseball/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/cricket/dream11-ipl-2020/final/49600/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/field_hockey/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/ice_hockey/whl-2019-2020/regular-season/34774/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/basketball/liga-endesa-2020-2021/regular-season/48319/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/water_polo/fina-world-league-2020/european-preliminaries-final/39428/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/field_hockey/belfius-eurohockey-championship-2019-antwerp/final/24481/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/kabaddi/vivo-pro-kabaddi-league-season-7/final/33143/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://datasportsgroup.com'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/rugby/top-14-2020-2021/regular-season/46912/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/leagueoflegends/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://twitter.com/DataSportsGroup'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/volleyball/superlega-credem-banca-2020-2021/regular-season/47984/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://twitter.com/glsportsarchive'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/handball/velux-ehf-champions-league-2019-2020/final/33352/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/ice_hockey/ohl-2019-2020/regular-season/34764/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/rugby/gallagher-premiership-rugby-2020-2021/regular-season/48965/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/basketball/7days-eurocup-2020-2021/top-16/46439/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/rugby/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/leagueoflegends/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/badminton/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/kabaddi/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/cricket/twenty20-internationals-2020/regular-round/37450/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/table_tennis/2020-03-08/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/handball/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/rugby/2021-03-26/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/volleyball/cev-champions-league-2020-2021/group-stage/48112/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/water_polo/european-water-polo-championship-2020-budapest/final/36671/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/fifa/2020-02-27/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/kabaddi/2019-10-19/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/tennis/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/field_hockey/2021-03-25/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/baseball/shinhan-bank-sol-kbo-league-2020/regular-season/44725/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competitions/cricket/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/tennis/-/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/volleyball/fivb-volleyball-womens-nations-league-2019/final/32194/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/water_polo/fina-world-cup-2018-berlin/final/24160/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/baseball/cpbl-2021/1st-phase/52234/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/handball/world-mens-handball-championship-2021-egypt/group-stage/48352/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/field_hockey/rabobank-hockey-champions-trophy-2018-netherlands/final/22623/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/badminton/2019-12-01/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/sport/rocketleague/2021-03-14/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/handball/ehf-euro-2020-sweden-austria-norway/final/33528/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/ice_hockey/nhl-2020-2021/regular-season/50565/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.linkedin.com/company/data-sports-group'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/field_hockey/euro-hockey-league-2019-2020/quarter-finals/35950/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://globalsportsarchive.com/competition/rugby/guinness-six-nations-2020/final-round/33205/'
-      }
-    ]
-  },
-  'https://www.sambafoot.com/br':
-  {
-    'visited_links': 151,
-    'harvest_ratio': 0.052980132450331126,
-    'relevant_links': 8,
-    'thread': 'Thread-15',
-    'frontier': [
-      {
-        'weight': 90.0,
-        'link': 'https://www.sambafoot.com/br/jogadores/147-philippe-coutinho'
-      },
-      {
-        'weight': 90.0,
-        'link': 'https://www.sambafoot.com/br/jogadores/275-neymar'
-      },
-      {
-        'weight': 90.0,
-        'link': 'https://www.sambafoot.com/br/jogadores/279-alisson-becker'
-      },
-      {
-        'weight': 90.0,
-        'link': 'https://www.sambafoot.com/br/jogadores'
-      },
-      {
-        'weight': 90.0,
-        'link': 'https://www.sambafoot.com/br/jogadores/299-roberto-firmino'
-      },
-      {
-        'weight': 70.0,
-        'link': 'https://www.sambafoot.com/br/fotos-de-futebol-bresil-venezuela-eliminatoires2018'
-      },
-      {
-        'weight': 70.0,
-        'link': 'https://www.sambafoot.com/br/fotos-de-futebol-argentine-bresil-novembre2015'
-      },
-      {
-        'weight': 70.0,
-        'link': 'https://www.sambafoot.com/br/fotos-de-futebol-bresil-perou-novembre2015'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2454-juventude'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2489-brusque'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/504-ponte-preta'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/inter-goleia-ypiranga-no-gauchao'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/en/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2490-ituano'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/arsenal-derrota-olympiacos-na-europa-league'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2459-oeste'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/97-atleticomg'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/93-flamengo'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/palmeiras-derrota-ferroviaria-no-paulista'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2487-manaus-fc'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/inter-derrota-novo-hamburgo-no-gauchao'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/aniversarios'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/contacto'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/videos-samba-de-ouro-2015'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/100-botafogo'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/fr/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.youtube.com/user/sambafoot'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/jogos'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/estadios'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2475-botafogo-pb'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/102-america-mineiro'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/94-cruzeiro'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.instagram.com/sambafootcom/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/99-parana'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2488-jacuipense'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/selecao'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/treinadores'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2074-bragantino'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2476-ferroviario'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/95-sport-recife'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/forum'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/897-goias'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/96-sao-paulo'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/91-chapecoensesc'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/104-atletico-paranaense'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/85-fluminense'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/978-nautico-recife'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/89-palmeiras'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/86-internacional'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2466-confianca'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/979-guarani-campinas'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2469-imperatriz'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2477-treze'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/593-figueirense'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/502-coritiba'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/fluminense-derrota-flamengo-no-campeonato-carioca'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/503-avai'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2453-paysandu'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/87-gremio'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2463-botafogo-sp'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2474-remo'
-      },
-      {
-        'weight': 0,
-        'link': 'https://twitter.com/Sambafoot'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/rb-bragantino-amassa-mirassol-na-copa-do-brasil'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/501-atletico-goianiense'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/sobre'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/gremio-confirma-classificacao-diante-ayacucho-na-libertadores'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/corinthians-vence-sao-caetano-no-paulistao'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/90-vasco-da-gama'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/92-corinthians'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/101-bahia'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/videos-samba-de-ouro-2011'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/103-ceara'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/samba-de-ouro'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2455-brasil-de-pelotas'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/592-santa-cruz'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/896-csa'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/98-santos'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2452-crb'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2465-operariopr'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/com-show-de-brasuccas-ajax-avanca-da-europa-league'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/noticias/flamengo-goleia-resende-no-campeonato-carioca'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/es/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/apostas/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/videos'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/88-vitoria'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2451-sampaio-correa'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/898-fortaleza-ec'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2464-cuiaba'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/videos-samba-de-ouro-2013'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.facebook.com/sambafoot/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/videos-samba-de-ouro-2014'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.linkedin.com/company/sambafoot/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sambafoot.com/br/times/2457-vila-nova'
-      }
-    ]
-  },
-  'https://www.transfermarkt.com.br':
-  {
-    'visited_links': 5,
-    'harvest_ratio': 0.0,
-    'relevant_links': 0,
-    'thread': 'Thread-13',
-    'frontier': [
-      {
-        'weight': 0,
-        'link': 'https://www.transfermarkt.com.br/as-50-maiores-lendas-na-historia-do-transfermarkt/index/galerie/1019'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.facebook.com/Transfermarkt.pt'
-      },
-      {
-        'weight': 0,
-        'link': 'http://instagram.com/transfermarkt_official'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.transfermarkt.com.br/campeonato-brasileiro-serie-a/marktwertaenderungen/wettbewerb/BRA1/pos//detailpos/0/verein_id/0/land_id/0/plus/1'
-      },
-      {
-        'weight': 0,
-        'link': 'https://twitter.com/TMpt_news'
-      }
-    ]
-  },
-  'https://fbref.com':
-  {
-    'visited_links': 75,
-    'harvest_ratio': 0.08,
-    'relevant_links': 6,
-    'thread': 'Thread-5',
-    'frontier': [
-      {
-        'weight': 80.0,
-        'link': 'https://www.sports-reference.com/blog/2021/02/fbref-scouting-reports-and-similar-players-launched/'
-      },
-      {
-        'weight': 80.0,
-        'link': 'https://fbref.com/en/stathead/player_comparison.cgi'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://statsbomb.com/'
-      },
-      {
-        'weight': 50.0,
-        'link': 'https://www.pro-football-reference.com/'
-      },
-      {
-        'weight': 40.0,
-        'link': 'https://stathead.com/profile/?do=logout'
-      },
-      {
-        'weight': 40.0,
-        'link': 'https://stathead.com/profile/?utm_source=web&utm_medium=fb&utm_campaign=sr-nav-bar-top-account'
-      },
-      {
-        'weight': 0,
-        'link': 'https://stathead.com/users/login.cgi?token=1'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/this-month.html'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/feed/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/en/linker/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.paypal.me/sportsref'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/2016/06/redesign-team-and-league-logos-courtesy-sportslogos-net/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.linkedin.com/company/sports-reference-llc'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/ways-sports-reference-can-help-your-website/?utm_medium=sr&utm_source=fb&utm_campaign=site-footer-ways-help'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/de/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/this-month.html?&utm_medium=sr&utm_source=fb&utm_campaign=site-footer-twisr'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/2021/01/katie-sharp-joins-sports-reference/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/es/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://stathead.com/?&utm_medium=sr&utm_source=fb&utm_campaign=site-footer'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/en/about/email'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/en/about/scouting-reports-explained'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/'
-      },
-      {
-        'weight': 0,
-        'link': 'http://datasportsgroup.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.youtube.com/user/sportsreference'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/pt/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/2018/06/fbref-com-launches-today'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/2018/06/fbref-com-launches-today/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.baseball-reference.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://stathead.com/?utm_source=web&utm_medium=fb&utm_campaign=sr-nav-bar-top-link'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/2020/11/fbref-adds-head-to-head-records/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/it/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.basketball-reference.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/en/about/coverage'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/en/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/category/stathead-tutorial-series/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://reddit.com/r/SportsReference'
-      },
-      {
-        'weight': 0,
-        'link': 'https://stathead.com/users/signup.cgi'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.hockey-reference.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/blog/sports-reference-social-media/'
-      },
-      {
-        'weight': 0,
-        'link': 'http://sportslogos.net/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/fr/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://twitter.com/FBref'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/cfb/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://widgets.sports-reference.com/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://fbref.com/en/about/'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.sports-reference.com/cbb/'
-      }
-    ]
-  },
-  'https://br.soccerway.com':
-  {
-    'visited_links': 30,
-    'harvest_ratio': 1.0,
-    'relevant_links': 30,
-    'thread': 'Thread-7',
-    'frontier': [
-      {
-        'weight': 60.0,
-        'link': 'https://es.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://cn.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://ru.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://br.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://nl.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://it.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://tr.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://au.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://pl.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://uk.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://gr.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://nr.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://za.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://int.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://ro.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://de.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://sg.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://us.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://ca.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://my.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://kr.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://ie.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://el.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://pt.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://gh.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://th.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://id.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://fr.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://ng.soccerway.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://ke.soccerway.com'
-      }
-    ]
-  },
-  'https://www.playmakerstats.com':
-  {
-    'visited_links': 80,
-    'harvest_ratio': 0.05,
-    'relevant_links': 4,
-    'thread': 'Thread-3',
-    'frontier': [
-      {
-        'weight': 60.0,
-        'link': 'https://www.soccerzz.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://www.playmakerstats.com'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://www.instagram.com/playmakerstats'
-      },
-      {
-        'weight': 60.0,
-        'link': 'https://facebook.com/playmakerstatsEN'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533127708~4.50&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1549620692~4.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125681~101.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.calciozz.it'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1550880909~2.25&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124374~2.80&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533126729~1.14&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124756~41.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125427~1.60&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125429~6.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1550880910~3.20&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124420~1.22&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125007~5.50&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125676~1.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125476~1.72&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533453018~3.60&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1549620690~1.80&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533453041~1.33&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124273~4.75&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533453085~8.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124232~1.28&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533126902~5.50&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125193~1.12&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.youtube.com/channel/UCNbC0JvCRvd7NUJAZbaxhvg'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125190~19.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1550880911~2.90&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124375~3.10&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124235~5.25&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533453017~4.75&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.ogol.com.br'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1549620691~3.50&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://twitter.com/playmaker_EN'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533126903~15.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125013~3.20&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124762~1.02&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124276~3.40&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124446~13.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533453087~1.36&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/dl/~offer?affiliate=365_033182'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125014~1.75&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125353~41.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533453039~11.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.zerozero.pt'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124342~13.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124339~26.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124277~1.80&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124236~10.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125428~3.60&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125351~15.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533453040~4.50&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533127709~9.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124761~17.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125474~5.25&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124445~6.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533126734~17.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124343~1.05&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533124376~2.62&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125475~3.40&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125191~8.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.leballonrond.fr'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533453086~4.75&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533127705~1.28&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.fussballzz.de'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125340~1.03&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533453019~1.75&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533126897~1.22&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533125680~51.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.bet365.com/betslip/instantbet/default.aspx?bs=1533126732~8.00&Instantbet=1&affiliatecode=365_505393'
-      },
-      {
-        'weight': 0,
-        'link': 'https://www.ceroacero.es'
-      }
-    ]
-  }
-}
