+++ wrapper/main.py
@@ -1,0 +1,4 @@
+from wrapper import Wrapper
+
+if __name__ == '__main__':
+    Wrapper.process()
